chore: move create relay logic to mixin
